+++ package.json
@@ -42,47 +42,47 @@
     "qs": "6.10.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "12.1.4",
-    "@commitlint/config-conventional": "12.1.4",
-    "@eth-optimism/smock": "1.1.9",
+    "@commitlint/cli": "13.1.0",
+    "@commitlint/config-conventional": "13.1.0",
+    "@eth-optimism/smock": "1.1.10",
     "@lbertenasco/contract-utils": "0.3.8",
     "@nomiclabs/hardhat-ethers": "npm:hardhat-deploy-ethers@0.3.0-beta.10",
     "@nomiclabs/hardhat-etherscan": "2.1.4",
     "@nomiclabs/hardhat-waffle": "2.0.1",
-    "@types/chai": "4.2.19",
+    "@types/chai": "4.2.21",
     "@types/chai-as-promised": "7.1.4",
-    "@types/lodash": "4.14.170",
-    "@types/mocha": "8.2.2",
+    "@types/lodash": "4.14.172",
+    "@types/mocha": "9.0.0",
     "@types/moment": "2.13.0",
-    "@types/node": "^16.4.6",
+    "@types/node": "^16.4.12",
     "@uniswap/v2-core": "1.0.1",
     "@uniswap/v2-periphery": "1.1.0-beta.0",
     "@uniswap/v3-periphery": "1.1.1",
-    "aws-sdk": "2.939.0",
+    "aws-sdk": "2.961.0",
     "axios": "0.21.1",
     "axios-retry": "3.1.9",
     "chai": "4.3.4",
     "chai-as-promised": "7.1.1",
     "dotenv": "10.0.0",
     "ethereum-waffle": "3.4.0",
-    "ethers": "5.4.1",
-    "hardhat": "2.4.3",
+    "ethers": "5.4.3",
+    "hardhat": "2.5.0",
     "hardhat-deploy": "0.8.11",
     "hardhat-gas-reporter": "1.0.4",
     "hardhat-preprocessor": "0.1.4",
-    "husky": "7.0.0",
-    "inquirer": "8.1.1",
+    "husky": "7.0.1",
+    "inquirer": "8.1.2",
     "lodash": "4.17.21",
     "moment": "2.29.1",
     "pinst": "2.1.6",
     "prettier": "2.3.2",
-    "prettier-plugin-solidity": "1.0.0-beta.13",
+    "prettier-plugin-solidity": "1.0.0-beta.17",
     "solhint": "3.3.6",
     "solhint-plugin-prettier": "0.0.5",
     "solidity-coverage": "0.7.16",
     "sort-package-json": "1.50.0",
-    "standard-version": "9.3.0",
-    "ts-node": "10.0.0",
+    "standard-version": "9.3.1",
+    "ts-node": "10.1.0",
     "typescript": "4.3.5"
   },
   "authors": [
